Move hingeRange, listToSlice, parseSlice into new module collection
